requirements.txt
streamlit
Pillow
numpy
tensorflow
wikipedia
google-generativeai
gdown